+++ lib/obsidian/parser.rb
@@ -2,9 +2,9 @@
 
 require_relative "parser/version"
 require_relative "parser/parsed_markdown_document"
-require_relative "parser/obsidian_flavored_markdown"
+require_relative "parser/markdown_parser"
 require_relative "parser/page"
-require_relative "parser/markdown_content"
+require_relative "parser/markdown_document"
 require_relative "parser/html_renderer"
 
 require "forwardable"
@@ -21,7 +21,7 @@ module Obsidian
 
     def initialize(vault_directory)
       @index = Obsidian::Page.create_root
-      renderer = HtmlRenderer.new
+      markdown_parser = MarkdownParser.new
 
       vault_directory.glob("**/*.md").each do |path|
         dirname, basename = path.relative_path_from(vault_directory).split
@@ -41,7 +41,7 @@ module Obsidian
         @index.add_page(
           slug,
           last_modified: path.mtime,
-          content: MarkdownContent.new(path, @index, renderer: renderer)
+          content: MarkdownDocument.new(path, @index, markdown_parser: markdown_parser)
         )
       end